chore: 🐝 Update SDK - Generate 0.1.1

+++ gusto_app_int/src/gusto_app_integration/models/get_v1_companies_company_id_contractor_payment_groupsop.py
@@ -21,9 +21,9 @@ class GetV1CompaniesCompanyIDContractorPaymentGroupsRequestTypedDict(TypedDict):
     r"""The time period for which to retrieve contractor payment groups. Defaults to 6 months ago."""
     end_date: NotRequired[str]
     r"""The time period for which to retrieve contractor payment groups. Defaults to today's date."""
-    page: NotRequired[float]
+    page: NotRequired[int]
     r"""The page that is requested. When unspecified, will load all objects unless endpoint forces pagination."""
-    per: NotRequired[float]
+    per: NotRequired[int]
     r"""Number of objects per page. For majority of endpoints will default to 25"""
     x_gusto_api_version: NotRequired[VersionHeader]
     r"""Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used."""
@@ -48,13 +48,13 @@ class GetV1CompaniesCompanyIDContractorPaymentGroupsRequest(BaseModel):
     r"""The time period for which to retrieve contractor payment groups. Defaults to today's date."""
 
     page: Annotated[
-        Optional[float],
+        Optional[int],
         FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
     ] = None
     r"""The page that is requested. When unspecified, will load all objects unless endpoint forces pagination."""
 
     per: Annotated[
-        Optional[float],
+        Optional[int],
         FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
     ] = None
     r"""Number of objects per page. For majority of endpoints will default to 25"""
